orbitview/requirements.txt
asgiref==3.8.1
django==5.1.7
djangorestframework==3.15.2
dnspython==2.7.0
mongoengine==0.29.1
pip==25.0
pymongo==3.11.4
python-dotenv==1.0.1
pytz==2025.1
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1

mysqlclient
pillow
